Vello CPU: feature flag filters
