+++ crates/anyrender_vello_cpu/Cargo.toml
@@ -22,6 +22,7 @@ log_frame_times = [
 # Useful for benchmarking (as we expect to eventually eliminate this cost)
 # but not recommended for use in production
 experimental_image_cache = []
+filters = []
 
 [dependencies]
 anyrender = { workspace = true }
