+++ crates/anyrender_vello_cpu/src/lib.rs
@@ -5,6 +5,9 @@ mod image_renderer;
 mod scene;
 mod window_renderer;
 
+#[cfg(feature = "filters")]
+mod filters;
+
 pub use image_renderer::VelloCpuImageRenderer;
 pub use scene::VelloCpuScenePainter;
